+++ utils/resources/models/certification/smooth_vit/head_bias.npy
