+++ utils/resources/models/certification/smooth_vit/head_weight.npy
